Базовые CRUD-операции с пользователями и объявлениями (#16)
* Поправил ошибку

* CRUD-методы для пользователей и объявлений

* Поправил ошибку в схеме пользователя

* Исправил ошибки аутентификации

* Поменял тестовую конфигурацию

* Правки схемы объявлений и базовых CRUD операций

* Тесты на базовые CRUD-операции

+++ backend/app/crud/base.py
@@ -61,6 +61,8 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
     def remove(self, db: Session, *, id: int) -> ModelType:
         obj = db.query(self.model).get(id)
+        if obj is None:
+            raise ValueError
         db.delete(obj)
         db.commit()
         return obj
